+++ 1.21.10/Common/src/main/java/fuzs/universalenchants/UniversalEnchants.java
@@ -2,6 +2,7 @@ package fuzs.universalenchants;
 
 import fuzs.puzzleslib.api.config.v3.ConfigHolder;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
+import fuzs.puzzleslib.api.core.v1.context.PackRepositorySourcesContext;
 import fuzs.puzzleslib.api.core.v1.utility.ResourceLocationHelper;
 import fuzs.puzzleslib.api.event.v1.FinalizeItemComponentsCallback;
 import fuzs.puzzleslib.api.event.v1.core.EventPhase;
@@ -12,6 +13,7 @@ import fuzs.universalenchants.config.ServerConfig;
 import fuzs.universalenchants.handler.BetterEnchantsHandler;
 import fuzs.universalenchants.handler.ItemCompatHandler;
 import fuzs.universalenchants.init.ModRegistry;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -22,6 +24,9 @@ public class UniversalEnchants implements ModConstructor {
     public static final Logger LOGGER = LogManager.getLogger(MOD_NAME);
 
     public static final ConfigHolder CONFIG = ConfigHolder.builder(MOD_ID).server(ServerConfig.class);
+    public static final ResourceLocation COMPATIBLE_DAMAGE_ENCHANTMENTS_LOCATION = id("compatible_damage_enchantments");
+    public static final ResourceLocation COMPATIBLE_PROTECTION_ENCHANTMENTS_LOCATION = id(
+            "compatible_protection_enchantments");
 
     @Override
     public void onConstructMod() {
@@ -39,6 +44,16 @@ public class UniversalEnchants implements ModConstructor {
         // run after other mods had a chance to change looting level
         LivingExperienceDropCallback.EVENT.register(EventPhase.AFTER, BetterEnchantsHandler::onLivingExperienceDrop);
         BlockEvents.DROP_EXPERIENCE.register(EventPhase.AFTER, BetterEnchantsHandler::onDropExperience);
+    }
+
+    @Override
+    public void onAddDataPackFinders(PackRepositorySourcesContext context) {
+        context.registerBuiltInPack(COMPATIBLE_DAMAGE_ENCHANTMENTS_LOCATION,
+                Component.literal("Compatible Damage Enchantments"),
+                false);
+        context.registerBuiltInPack(COMPATIBLE_PROTECTION_ENCHANTMENTS_LOCATION,
+                Component.literal("Compatible Protection Enchantments"),
+                false);
     }
 
     public static ResourceLocation id(String path) {