+++ 1.21.10/NeoForge/src/main/java/fuzs/universalenchants/neoforge/UniversalEnchantsNeoForge.java
@@ -3,12 +3,13 @@ package fuzs.universalenchants.neoforge;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
 import fuzs.puzzleslib.neoforge.api.data.v2.core.DataProviderHelper;
 import fuzs.universalenchants.UniversalEnchants;
-import fuzs.universalenchants.data.tags.ModBlockTagProvider;
 import fuzs.universalenchants.data.ModDatapackRegistriesProvider;
-import fuzs.universalenchants.data.tags.ModEnchantmentTagProvider;
-import fuzs.universalenchants.data.tags.ModItemTagProvider;
+import fuzs.universalenchants.data.tags.ModBlockTagsProvider;
+import fuzs.universalenchants.data.tags.ModEnchantmentTagsProvider;
+import fuzs.universalenchants.data.tags.ModItemTagsProvider;
 import fuzs.universalenchants.handler.BetterEnchantsHandler;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.packs.PackType;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.common.NeoForge;
@@ -22,9 +23,15 @@ public class UniversalEnchantsNeoForge {
         registerEventHandlers(NeoForge.EVENT_BUS);
         DataProviderHelper.registerDataProviders(UniversalEnchants.MOD_ID,
                 ModDatapackRegistriesProvider::new,
-                ModItemTagProvider::new,
-                ModBlockTagProvider::new,
-                ModEnchantmentTagProvider::new);
+                ModItemTagsProvider::new,
+                ModBlockTagsProvider::new,
+                ModEnchantmentTagsProvider::new);
+        DataProviderHelper.registerDataProviders(UniversalEnchants.COMPATIBLE_DAMAGE_ENCHANTMENTS_LOCATION,
+                PackType.SERVER_DATA,
+                ModEnchantmentTagsProvider.DamageEnchantments::new);
+        DataProviderHelper.registerDataProviders(UniversalEnchants.COMPATIBLE_PROTECTION_ENCHANTMENTS_LOCATION,
+                PackType.SERVER_DATA,
+                ModEnchantmentTagsProvider.ProtectionEnchantments::new);
     }
 
     private static void registerEventHandlers(IEventBus eventBus) {