+++ 1.21.10/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.10.1-1.21.10] - 2025-10-23
+
+### Changed
+
+- All protection and damage enchantments are once again incompatible by default with each other
+- Previously, sharpness was compatible with more specialized damage enchantments like smite, impaling, etc.; the same
+  for protection
+- Restoring the old behavior is possible from two built-in data packs now, which can be enabled via `/datapack` or when
+  creating a new world
+
 ## [v21.10.0-1.21.10] - 2025-10-08
 
 ### Changed